requirements.txt
selenium
python-dotenv
pillow
pytesseract
huggingface_hub